# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "local_bridge",
    "description": "Media AI task queue bridge with OpenAPI docs",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/single/first-frame-image": {
      "post": {
        "tags": [
          "single",
          "single"
        ],
        "summary": "Create First Frame Image",
        "operationId": "create_first_frame_image_v1_single_first_frame_image_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FirstFrameImageCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FirstFrameImageCreatedResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response 400 Create First Frame Image V1 Single First Frame Image Post"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response 404 Create First Frame Image V1 Single First Frame Image Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/single/jimeng-image": {
      "post": {
        "tags": [
          "single",
          "single"
        ],
        "summary": "Create Jimeng Image",
        "operationId": "create_jimeng_image_v1_single_jimeng_image_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JimengImageCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JimengImageCreatedResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response 400 Create Jimeng Image V1 Single Jimeng Image Post"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response 404 Create Jimeng Image V1 Single Jimeng Image Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/single/jimeng-video": {
      "post": {
        "tags": [
          "single",
          "single"
        ],
        "summary": "Create Jimeng Video",
        "operationId": "create_jimeng_video_v1_single_jimeng_video_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JimengVideoCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JimengVideoCreatedResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response 400 Create Jimeng Video V1 Single Jimeng Video Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FirstFrameImageCreateRequest": {
        "properties": {
          "styleImageId": {
            "type": "string",
            "title": "Styleimageid"
          },
          "sceneId": {
            "type": "string",
            "title": "Sceneid"
          },
          "force": {
            "type": "boolean",
            "title": "Force",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "styleImageId",
          "sceneId"
        ],
        "title": "FirstFrameImageCreateRequest"
      },
      "FirstFrameImageCreatedResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "dryRun": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dryrun"
          },
          "caseFile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Casefile"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "job": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GptFirstFrameImageJob"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "ok"
        ],
        "title": "FirstFrameImageCreatedResponse"
      },
      "GptFirstFrameImageJob": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "caseFile": {
            "type": "string",
            "title": "Casefile"
          },
          "mediaAi": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GptFirstFrameImageMediaAi"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "caseFile"
        ],
        "title": "GptFirstFrameImageJob"
      },
      "GptFirstFrameImageMediaAi": {
        "properties": {
          "kind": {
            "type": "string",
            "title": "Kind",
            "default": "first-frame-image"
          },
          "platform": {
            "type": "string",
            "title": "Platform",
            "default": "gpt"
          },
          "baseUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseurl"
          },
          "productId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Productid"
          },
          "productName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Productname"
          },
          "ipId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ipid"
          },
          "styleImageId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Styleimageid"
          },
          "sceneId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sceneid"
          },
          "uploadSubDir": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uploadsubdir"
          }
        },
        "type": "object",
        "title": "GptFirstFrameImageMediaAi",
        "description": "mediaAi for /v1/single/first-frame-image."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JimengImageCreateRequest": {
        "properties": {
          "styleImageId": {
            "type": "string",
            "title": "Styleimageid"
          },
          "sceneId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sceneid"
          },
          "productId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Productid"
          },
          "ipId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ipid"
          },
          "force": {
            "type": "boolean",
            "title": "Force",
            "default": false
          },
          "prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt"
          },
          "noEmbedCookie": {
            "type": "boolean",
            "title": "Noembedcookie",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "styleImageId"
        ],
        "title": "JimengImageCreateRequest"
      },
      "JimengImageCreatedResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "dryRun": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dryrun"
          },
          "caseFile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Casefile"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "job": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JimengImageJob"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "ok"
        ],
        "title": "JimengImageCreatedResponse"
      },
      "JimengImageJob": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "caseFile": {
            "type": "string",
            "title": "Casefile"
          },
          "mediaAi": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JimengImageMediaAi"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "caseFile"
        ],
        "title": "JimengImageJob"
      },
...
      "JimengVideoCreateRequest": {
        "properties": {
          "productId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Productid"
          },
          "ipId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ipid"
          },
          "firstFrameId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Firstframeid"
          },
          "movementId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Movementid"
          },
          "force": {
            "type": "boolean",
            "title": "Force",
            "default": false
          },
          "prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt"
          },
          "noEmbedCookie": {
            "type": "boolean",
            "title": "Noembedcookie",
            "default": false
          }
        },
        "type": "object",
        "title": "JimengVideoCreateRequest"
      },
      "JimengVideoCreatedResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "dryRun": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dryrun"
          },
          "caseFile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Casefile"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "job": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JimengVideoJob"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "ok"
        ],
        "title": "JimengVideoCreatedResponse"
      },
      "JimengVideoJob": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "caseFile": {
            "type": "string",
            "title": "Casefile"
          },
          "mediaAi": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JimengVideoMediaAi"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "caseFile"
        ],
        "title": "JimengVideoJob"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}